# Conflicts:
#	.idea/misc.xml
#	CMakeLists.txt

+++ maths/Matrix.h
@@ -11,7 +11,7 @@
 
 class Matrix {
 private:
-    double matrix[4][4];
+    float matrix[4][4];
 public:
     Matrix();
     ~Matrix();
@@ -21,43 +21,43 @@ public:
     bool isIdentity();
 
     /** Getter and setter operations */
-    double get(int row, int column) const {return matrix[row][column];}
-    void set(int row, int column, double value){matrix[row][column] = value;}
+    float get(int row, int column) const {return matrix[row][column];}
+    void set(int row, int column, float value){matrix[row][column] = value;}
 
-    void setPoint(double x, double y, double z);
+    void setPoint(float x, float y, float z);
     Vector getPoint() const;
-    void setVector(double x, double y, double z, int positionVector);
+    void setVector(float x, float y, float z, int positionVector);
     void setVector(Vector &v, int positionVector);
     Vector getVector(int positionVector) const;
-    void setScale(double sx, double sy, double sz);
-    void setRotX(double x);
-    void setRotY(double y);
-    void setRotZ(double z);
+    void setScale(float sx, float sy, float sz);
+    void setRotX(float x);
+    void setRotY(float y);
+    void setRotZ(float z);
 
     /** Return types of matrices **/
     Matrix transpose() const;
     Matrix adjugate() const;
     Matrix inverse() const;
-    double determinant() const;
+    float determinant() const;
 
     /** Definition of special matrices **/
-    static Matrix traslation(double x, double y, double z);
-    static Matrix scale(double sx, double sy, double sz);
-    static Matrix rotationXYZ(double x, double y, double z);
-    static double determinant3x3(double a11, double a12, double a13,
-                                 double a21, double a22, double a23,
-                                 double a31, double a32, double a33);
+    static Matrix traslation(float x, float y, float z);
+    static Matrix scale(float sx, float sy, float sz);
+    static Matrix rotationXYZ(float x, float y, float z);
+    static float determinant3x3(float a11, float a12, float a13,
+                                 float a21, float a22, float a23,
+                                 float a31, float a32, float a33);
 
     /** Operator overloading **/
     Matrix operator+(const Matrix &m) const;
     Matrix operator-(const Matrix &m) const;
-    Matrix operator*(double m) const;
+    Matrix operator*(float m) const;
     Matrix operator*(const Matrix &m) const;
     Vector operator*(const Vector &pd) const;
 
     Matrix operator+=(const Matrix &m);
     Matrix operator-=(const Matrix &m);
-    Matrix operator*=(double m);
+    Matrix operator*=(float m);
     Matrix operator*=(const Matrix &m);
 
     /** Output screen **/
